split example-specific URI generation into its own class
